Reenable EcAdapterInterface functions as upstream secp256k1 now supports sign_compact / recover_compact Added test for compact signatures
