+++ tests/AbstractTestCase.php
@@ -16,9 +16,9 @@ class AbstractTestCase extends \PHPUnit_Framework_TestCase
         $math = Bitcoin::getMath();
         $generator = Bitcoin::getGenerator();
 
-        $adapters = array(
+        $adapters = [
             [new PhpEcc($math, $generator)]
-        );
+        ];
 
         if (!defined('HHVM_VERSION') && extension_loaded('secp256k1')) {
             $adapters[] = [new Secp256k1($math, $generator)];
